```
❯ go test; go vet; golint
PASS
ok      _/Users/andy/dev/durafmt        0.007s
```

Resolves #16