@adharmawan 2019-10-06
Hi @hako, could you take a look? I know the coverage decreased (by a bit). However, I think it's better than ignoring the error