@hako 2019-10-06
@adharmawan LGTM, did a `benchcmp` of this and it is faster.

```
benchcmp old.txt new.txt
benchmark                       old ns/op     new ns/op     delta
BenchmarkParse-4                223           117           -47.53%
BenchmarkParseStringShort-4     436           395           -9.40%
BenchmarkParseString-4          409           397           -2.93%
```